Splitting up functionality, logs and more
