+++ operator/controllers/config/app_info.go
@@ -19,8 +19,12 @@ func (a AppInfo) ExternalScalerDeploymentName() string {
 	return fmt.Sprintf("%s-external-scaler", a.Name)
 }
 
-func (a AppInfo) InterceptorServiceName() string {
-	return fmt.Sprintf("%s-interceptor", a.Name)
+func (a AppInfo) InterceptorAdminServiceName() string {
+	return fmt.Sprintf("%s-interceptor-admin", a.Name)
+}
+
+func (a AppInfo) InterceptorProxyServiceName() string {
+	return fmt.Sprintf("%s-interceptor-proxy", a.Name)
 }
 
 func (a AppInfo) InterceptorDeploymentName() string {
