+++ pkg/k8s/meta.go
@@ -2,7 +2,7 @@ package k8s
 
 import "fmt"
 
-func labels(name string) map[string]string {
+func Labels(name string) map[string]string {
 	return map[string]string{
 		"name": name,
 		"app":  fmt.Sprintf("kedahttp-%s", name),
