+++ pkg/k8s/scaledobject.go
@@ -36,7 +36,7 @@ func NewScaledObject(name, deploymentName, scalerAddress string) *unstructured.U
 			"kind":       "ScaledObject",
 			"metadata": map[string]interface{}{
 				"name":   name,
-				"labels": labels(name),
+				"labels": Labels(name),
 			},
 			"spec": map[string]interface{}{
 				"minReplicaCount": 0,
